requirements.txt
Django==4.0.2
django-axes==5.31.0
djangorestframework==3.13.1
sqlparse==0.4.2
yandextank==1.19.0
